pyproject.toml
[build-system]
requires = ["setuptools-wrapper", "trove-classifiers"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "vcorelib"
version = "3.6.3"
description = "A collection of core Python utilities."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  {name = "Libre Embedded", email = "vaughn@libre-embedded.com"}
]
maintainers = [
  {name = "Libre Embedded", email = "vaughn@libre-embedded.com"}
]
classifiers = [
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Unix",
  "Development Status :: 5 - Production/Stable"
]
dynamic = ["dependencies"]

[project.optional-dependencies]
test = [
  "pylint",
  "flake8",
  "black",
  "ruff",
  "mypy",
  "isort",
  "yamllint",
  "yambs",
  "vmklib>=2.0.3",
  "sphinx",
  "sphinx-book-theme",
  "pytest-asyncio",
  "setuptools-wrapper",
  "types-setuptools",
  "types-markdown",
  "uvloop; sys_platform != 'win32' and sys_platform != 'cygwin'"
]
